+++ src/actions/mod.rs
@@ -2,4 +2,5 @@ pub mod create_task_definition;
 pub mod delete_task_definition;
 pub mod list_task_definition;
 pub mod patch_task_definition;
+pub mod run_pending_job;
 pub mod submit_job;
