+++ src/background/runner.rs
@@ -4,8 +4,8 @@ use sea_orm::{
 };
 
 use crate::{
+    actions::run_pending_job::run_pending_job,
     db::entities::{self, job::JobStatus},
-    docker::run_container,
 };
 
 pub async fn start_runner_loop(database_connection: DatabaseConnection) {
@@ -33,7 +33,7 @@ pub async fn start_runner_loop(database_connection: DatabaseConnection) {
             }
 
             for pending_job in pending_jobs {
-                if let Err(error) = process_pending_job(&database_connection, &pending_job).await {
+                if let Err(error) = run_pending_job(&database_connection, &pending_job).await {
                     println!("Error processing job: {:?}", error);
                     let mut job_active_model = pending_job.into_active_model();
                     job_active_model.status = Set(JobStatus::Failed);
@@ -46,36 +46,4 @@ pub async fn start_runner_loop(database_connection: DatabaseConnection) {
         }
     })
     .await;
-}
-
-async fn process_pending_job(
-    database_connection: &DatabaseConnection,
-    pending_job: &entities::job::Model,
-) -> anyhow::Result<()> {
-    // TODO: 리소스 제한이나 실행 제한 등에 걸리지 않는지 확인 (차후 개발)
-
-    // job 상태를 START로 변경
-    let mut job_active_model = pending_job.clone().into_active_model();
-    job_active_model.status = Set(JobStatus::Starting);
-    job_active_model.started_at = Set(Some(chrono::Utc::now()));
-    let pending_job = job_active_model.clone().update(database_connection).await?;
-
-    let Ok(Some(task_definition)) = entities::task_definition::Entity::find()
-        .filter(entities::task_definition::Column::Id.eq(pending_job.task_definition_id))
-        .one(database_connection)
-        .await
-    else {
-        return Err(anyhow::anyhow!("Error fetching task definition"));
-    };
-
-    // 컨테이너 실행
-    let container_id = run_container(task_definition)?;
-
-    // 컨테이너 정보를 job에 업데이트, job 상태를 RUNNING으로 변경
-    let mut job_active_model = pending_job.into_active_model();
-    job_active_model.container_id = Set(Some(container_id));
-    job_active_model.status = Set(JobStatus::Running);
-    job_active_model.update(database_connection).await?;
-
-    Ok(())
 }
